aws-community-day/requirements.txt
paho-mqtt
Adafruit_DHT
python-dotenv
python-decouple